fix(build): HookRegistry API arity, sign-conversions, size_t in benches
Four more compile errors surface as g++ 15 reaches the test layer
under -Werror=sign-conversion -Werror=conversion.

1) src/isolation/isolation_service.cpp:168
   HookRegistry::unregisterByModule has two overloads:
     - HookChain::unregisterByModule(ModuleId)            // single-chain
     - HookRegistry::unregisterByModule(HookPoint, ModuleId)
   The IsolationService caller used the 1-arg form, but we have a
   HookRegistry handle, not a HookChain — only the 2-arg form exists.
   Fixed by passing core::HookPoint::POST_FORK (the only point we
   register on, per audit fix O1).

2) tests/core/unit/test_capability_concurrency.cpp:79
   `static_cast<uint64_t>(id) * 0x9E3779B97F4A7C15ULL` — on LP64
   uint64_t is unsigned long while the literal is unsigned long long.
   Both are 64-bit but different types, and -Werror=sign-conversion
   flagged the implicit conversion of the product. Wrapped the
   literal in `static_cast<uint64_t>(...)` so both operands are the
   same type.

3) tests/core/bench_validate.cpp:111
   Same `pct * v.size()` size_t→double implicit conversion the
   earlier patch fixed in bench_harness.h. Caught now because
   bench_capability_validate doesn't include bench_harness.h —
   it has its own copy of the percentile lambda. Added the same
   explicit static_cast<double>(v.size()).

4) tests/core/unit/test_capability_survivability.cpp
   Multiple `int r`, `int c` loop variables used as std::vector
   subscripts:
     - tree[r], roots[r], tree[r][c]                         (6 sites)
     - lastTrue[idxFor(r,c)]                                  (2 sites)

   Strategy:
     a) Changed idxFor lambda to return std::size_t with explicit
        casts inside (handles the two lastTrue subscripts).
     b) Where r/c flow from a `for (int r=0; r<kRoots; ++r)` loop,
        bound a local `const auto ur = static_cast<std::size_t>(r)`
        and used that for the subscripts (cleaner than 4× inline
        static_cast).
     c) Where r/c come from uniform_int_distribution<int> (the
        producer thread's random subscript), used inline
        static_cast<std::size_t>() at the subscript sites — can't
        easily change the distribution to size_t because
        uniform_int_distribution<size_t> behaviour is platform-
        defined on LP64.

   Loop arithmetic that wasn't a subscript (e.g. `r * 100 + c` for
   the owner-id derivation) left as-is — no sign-conv there.

After pull:
  cd ~/projects/Astra
  git pull
  ./scripts/compile-astra.sh

+++ tests/core/bench_validate.cpp
@@ -108,7 +108,12 @@ static Stats fnSummarise(std::vector<uint64_t>& v, double fTscPerNs)
     if (v.empty()) return s;
     std::sort(v.begin(), v.end());
     auto fnAt = [&](double pct) {
-        size_t i = std::min<size_t>(static_cast<size_t>(pct * v.size()), v.size() - 1);
+        // Explicit double-cast on v.size() — Clang 21 / GCC 15 flag the
+        // implicit size_t → double conversion as potentially lossy under
+        // -Werror=conversion.
+        size_t i = std::min<size_t>(
+            static_cast<size_t>(pct * static_cast<double>(v.size())),
+            v.size() - 1);
         return static_cast<double>(v[i]) / fTscPerNs;
     };
     long double sum = 0;
